+++ jvm/selfie-lib/src/commonTest/kotlin/com/diffplug/selfie/guts/SourceFileToBeTest.kt
@@ -93,12 +93,7 @@ class SourceFileToBeTest {
     val functionCallPlusArg = functionCallPlusArgRaw.replace('\'', '"')
     val arg = argRaw.replace('\'', '"')
     val parsed = SourceFile("UnderTest.java", source)
-    if (source.contains(".toBe_TODO(")) {
-      parsed.parseToBe_TODO(1).functionCallPlusArg.toString() shouldBe functionCallPlusArg
-      parsed.parseToBe_TODO(1).arg.toString() shouldBe arg
-    } else {
-      parsed.parseToBe(1).functionCallPlusArg.toString() shouldBe functionCallPlusArg
-      parsed.parseToBe(1).arg.toString() shouldBe arg
-    }
+    parsed.parseToBeLike(1).functionCallPlusArg.toString() shouldBe functionCallPlusArg
+    parsed.parseToBeLike(1).arg.toString() shouldBe arg
   }
 }